+++ internal/cli/jira_issue.go
@@ -15,6 +15,7 @@ import (
 
 func jiraIssueCmd() *cobra.Command {
 	c := &cobra.Command{Use: "issue", Short: "Issue operations"}
+	c.AddCommand(jiraIssueTypesCmd(), jiraIssueCreateCheckCmd())
 
 	var fields string
 	get := &cobra.Command{